# RFC: Use `find` for listing files in subversion repositories, for a massive speedup