@bbatsov 2016-03-06
I don't like this approach, as it deviates from all other defaults. The command is configurable for a reason. I understand the default is slow, but I also think most users would expect their file list to be built using `svn` (especially if their not on a Unix OS).