@Wilfred 2016-03-07
Fair enough, thanks for your feedback. It's no problem for me to set this variable just within my environment.

I'll open a fresh PR if there's a flag to `svn ls` that prevents network access, but it looks like no such flag exists.