@Wilfred 2016-06-30
Turns out I set the default behaviour in #285! I'd totally forgotten. At the time some other users encountered similar issues (see the comments on 1304890963efd568b74f80622c2a6ab289d6f5d1).